# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ai-agent/sessions": {
      "post": {
        "operationId": "AiChatController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "title": {
                      "type": "string"
                    },
                    "modelProvider": {
                      "type": "string"
                    },
                    "modelName": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "closed"
                      ]
                    },
                    "version": {
                      "type": "integer"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an AI chat session",
        "tags": [
          "AI Agent"
        ]
      },
      "get": {
        "operationId": "AiChatController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Own user ID only; tenant admins may filter audit logs by another member",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped paginated records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "page",
                    "pageSize"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "tenantId": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "title": {
                            "type": "string"
                          },
                          "modelProvider": {
                            "type": "string"
                          },
                          "modelName": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "closed"
                            ]
                          },
                          "version": {
                            "type": "integer"
                          },
                          "messages": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "additionalProperties": true
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the current user chat sessions",
        "tags": [
          "AI Agent"
        ]
      }
    },
    "/api/ai-agent/sessions/{id}": {
      "patch": {
        "operationId": "AiChatController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "title": {
                      "type": "string"
                    },
                    "modelProvider": {
                      "type": "string"
                    },
                    "modelName": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "closed"
                      ]
                    },
                    "version": {
                      "type": "integer"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update own session; closing clears short-term memory",
        "tags": [
          "AI Agent"
        ]
      },
      "get": {
        "operationId": "AiChatController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "title": {
                      "type": "string"
                    },
                    "modelProvider": {
                      "type": "string"
                    },
                    "modelName": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "closed"
                      ]
                    },
                    "version": {
                      "type": "integer"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Load one chat session with message history",
        "tags": [
          "AI Agent"
        ]
      },
      "delete": {
        "operationId": "AiChatController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized record/result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete one chat session and its messages",
        "tags": [
          "AI Agent"
        ]
      }
    },
    "/api/ai-agent/chat": {
      "post": {
        "operationId": "AiChatController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User and assistant messages persisted"
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Send a message and receive a model response",
        "tags": [
          "AI Agent"
        ]
      }
    },
    "/api/ai-agent/files": {
      "post": {
        "operationId": "AiChatController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing, invalid, or expired JWT"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upload up to five chat attachments",
        "tags": [
          "AI Agent"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dashboard API",
    "description": "Authentication, business management, dashboard aggregation, tenant-isolated AI records, conversations and read-only AI tools.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "YearRangeDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "number",
            "example": 1950,
            "minimum": 1800,
            "maximum": 2100
          },
          "end": {
            "type": "number",
            "example": 2023,
            "minimum": 1800,
            "maximum": 2100
          }
        }
      },
      "NumberRangeDto": {
        "type": "object",
        "properties": {
          "min": {
            "type": "number",
            "example": 0,
            "minimum": 0
          },
          "max": {
            "type": "number",
            "example": 1000000,
            "minimum": 0
          }
        }
      },
      "CompanyBarChartFilterDto": {
        "type": "object",
        "properties": {
          "level": {
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "country": {
            "example": [
              "China",
              "USA"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "city": {
            "example": [
              "Beijing",
              "New York"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "founded_year": {
            "$ref": "#/components/schemas/YearRangeDto"
          },
          "annual_revenue": {
            "$ref": "#/components/schemas/NumberRangeDto"
          },
          "employees": {
            "$ref": "#/components/schemas/NumberRangeDto"
          }
        }
      },
      "BarChartQueryDto": {
        "type": "object",
        "properties": {
          "dimension": {
            "type": "string",
            "enum": [
              "level",
              "country",
              "city"
            ],
            "example": "country"
          },
          "filter": {
            "$ref": "#/components/schemas/CompanyBarChartFilterDto"
          }
        },
        "required": [
          "dimension",
          "filter"
        ]
      },
...
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Supply chain analysis"
          },
          "provider": {
            "type": "string",
            "enum": [
              "mock",
              "openai",
              "deepseek",
              "qwen",
              "kimi"
            ],
            "default": "mock"
          },
          "model": {
            "type": "string",
            "example": "deepseek-chat"
          }
        }
      },
      "UpdateSessionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Supply chain analysis"
          },
          "provider": {
            "type": "string",
            "enum": [
              "mock",
              "openai",
              "deepseek",
              "qwen",
              "kimi"
            ],
            "default": "mock"
          },
          "model": {
            "type": "string",
            "example": "deepseek-chat"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "closed"
            ]
          }
        }
      },
      "ChatAttachmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "mimeType",
          "size",
          "url"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "companyQuery": {
            "description": "Explicit read-only input for calling the existing company aggregation service",
            "allOf": [
              {
                "$ref": "#/components/schemas/BarChartQueryDto"
              }
            ]
          },
          "sessionId": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "maxLength": 10000
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatAttachmentDto"
            }
          }
        },
        "required": [
          "sessionId",
          "message"
        ]
      },
...
    }
  }
}